requirements.txt
flask
gunicorn
openai
yt-dlp
python-dotenv
